+++ project/agent/agent.h
@@ -22,8 +22,8 @@ class Agent
 
     static constexpr size_t NR_ACTIONS = 4; // Hardcoded number of actions
     size_t d_killedByAshTime = 500;
-    double Q_0 = 0;
-    double d_runReward;
+    float Q_0 = 0;
+    float d_runReward;
 
     size_t d_time = 0;
     SimContainer *d_maze;         // The maze the agent is navigating
@@ -32,27 +32,27 @@ class Agent
     size_t d_oldstate;
     size_t d_lastAction;
     size_t d_nrEpisodes;
-    std::vector<double> d_rewards;
+    std::vector<float> d_rewards;
     std::vector<size_t> d_hasDied;
 
-    double EPSILON;
+    float EPSILON;
 
     public:
-    Agent(size_t nrEpisodes, double epsilon);
+    Agent(size_t nrEpisodes, float epsilon);
     virtual ~Agent() = default;
 
     void run();
     bool performOneStep();
     void initialState(size_t state);
-    std::vector<double> & rewards();
+    std::vector<float> & rewards();
     std::vector<size_t> & hasDied();
     void maze(SimContainer *maze);
-    double runReward();
+    float runReward();
     virtual void stateSpaceSize(size_t size) = 0;
 
     protected:
     virtual Actions action(size_t stateIdx) = 0;
-    virtual void giveFeedback(double reward, size_t newStateIdx) = 0;
+    virtual void giveFeedback(float reward, size_t newStateIdx) = 0;
     virtual void newEpisode(size_t initialState);
 };
