+++ project/agent/qlearning/qlearning.h
@@ -6,16 +6,16 @@
 
 class QLearning : public Agent
 {
-  double d_gamma;
-  double d_alpha;
-  std::vector<std::vector<double>> d_QTable;
+  float d_gamma;
+  float d_alpha;
+  std::vector<std::vector<float>> d_QTable;
 
   public:
-    QLearning(size_t nrEpisodes, double discountFactor, double stepSize = 0.1, double epsilon = 0.1);
+    QLearning(size_t nrEpisodes, float discountFactor, float stepSize = 0.1, float epsilon = 0.1);
     ~QLearning();
 
     virtual Actions action(size_t stateIdx) override;
-    virtual void giveFeedback(double reward, size_t newState) override;
+    virtual void giveFeedback(float reward, size_t newState) override;
 
     virtual void stateSpaceSize(size_t size) override;
   private:
